requirements.txt
# LarkScout runtime dependencies
# Python >= 3.11 required

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Browser automation
playwright>=1.45.0

# Document parsing (MarkItDown by Microsoft)
markitdown[pdf,docx,pptx,xlsx,xls]>=0.1.0
pymupdf>=1.24.0

# LLM / OCR
google-genai>=0.8.0
openai>=1.0.0

# Image processing
Pillow>=10.4.0
numpy>=1.26.0,<2.0.0

# HTTP client
httpx>=0.27.0

# Test support
pytest-asyncio>=0.24.0
